first version of emitter framework in place

+++ routes/data.go
@@ -103,6 +103,11 @@ func DeleteDataByID(c *fiber.Ctx) error {
 	id := c.Params("id")
 	table := c.Params("type")
 	item := types.CreateType(table)
+	if item == nil {
+		err := fmt.Errorf("type not found in data type registry: %s", table)
+		db.Log("error", "Failed to create item", err.Error())
+		return c.Status(503).SendString(err.Error())
+	}
 
 	if err := db.DB.Delete(item, id).Error; err != nil {
 		db.Log("error", "Failed to delete item", err.Error())
